+++ src/main/java/com/cshome/toy/discopy/domain/member/dto/MemberInfoDto.java
@@ -1,6 +1,6 @@
 package com.cshome.toy.discopy.domain.member.dto;
 
-import com.cshome.toy.discopy.domain.member.Member;
+import com.cshome.toy.discopy.domain.member.entity.Member;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
