+++ src/main/java/com/cshome/toy/discopy/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.cshome.toy.discopy.domain.member.repository;
 
-import com.cshome.toy.discopy.domain.member.Member;
+import com.cshome.toy.discopy.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
