+++ src/main/java/com/cshome/toy/discopy/domain/member/service/CustomMemberDetailsService.java
@@ -1,24 +1,17 @@
 package com.cshome.toy.discopy.domain.member.service;
 
-import com.cshome.toy.discopy.domain.member.Member;
 import com.cshome.toy.discopy.domain.member.dto.MemberDetails;
 import com.cshome.toy.discopy.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-
-import java.util.Collections;
 
 @Service
 @RequiredArgsConstructor
 public class CustomMemberDetailsService implements UserDetailsService {
     private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
